cicd/pull_images/requirements.txt
docker==7.1.0
PyYAML==6.0.2
python-dotenv==1.1.0
